src/old/requirements.txt
tensorflow
pandas
numpy
sklearn
google-cloud-storage
scipy
tempfile
datetime
os